Move supla coordinator to coordinator module
