+++ homeassistant/components/supla/__init__.py
@@ -2,8 +2,6 @@
 
 from __future__ import annotations
 
-import asyncio
-from datetime import timedelta
 import logging
 
 from asyncpysupla import SuplaAPI
@@ -15,15 +13,14 @@ from homeassistant.helpers import config_validation as cv
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.discovery import async_load_platform
 from homeassistant.helpers.typing import ConfigType
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
+
+from .const import DOMAIN, SUPLA_COORDINATORS, SUPLA_SERVERS
+from .coordinator import SuplaCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
-DOMAIN = "supla"
 CONF_SERVER = "server"
 CONF_SERVERS = "servers"
-
-SCAN_INTERVAL = timedelta(seconds=10)
 
 SUPLA_FUNCTION_HA_CMP_MAP = {
     "CONTROLLINGTHEROLLERSHUTTER": Platform.COVER,
@@ -32,8 +29,7 @@ SUPLA_FUNCTION_HA_CMP_MAP = {
     "LIGHTSWITCH": Platform.SWITCH,
 }
 SUPLA_FUNCTION_NONE = "NONE"
-SUPLA_SERVERS = "supla_servers"
-SUPLA_COORDINATORS = "supla_coordinators"
+
 
 SERVER_CONFIG = vol.Schema(
     {
@@ -90,7 +86,7 @@ async def async_setup(hass: HomeAssistant, base_config: ConfigType) -> bool:
     return True
 
 
-async def discover_devices(hass, hass_config):
+async def discover_devices(hass: HomeAssistant, hass_config: ConfigType) -> None:
     """Run periodically to discover new devices.
 
     Currently it is only run at startup.
@@ -98,24 +94,7 @@ async def discover_devices(hass, hass_config):
     component_configs: dict[Platform, dict[str, dict]] = {}
 
     for server_name, server in hass.data[DOMAIN][SUPLA_SERVERS].items():
-
-        async def _fetch_channels():
-            async with asyncio.timeout(SCAN_INTERVAL.total_seconds()):
-                return {
-                    channel["id"]: channel
-                    for channel in await server.get_channels(  # noqa: B023
-                        include=["iodevice", "state", "connected"]
-                    )
-                }
-
-        coordinator = DataUpdateCoordinator(
-            hass,
-            _LOGGER,
-            name=f"{DOMAIN}-{server_name}",
-            update_method=_fetch_channels,
-            update_interval=SCAN_INTERVAL,
-        )
-
+        coordinator = SuplaCoordinator(hass, server, server_name)
         await coordinator.async_refresh()
 
         hass.data[DOMAIN][SUPLA_COORDINATORS][server_name] = coordinator
