+++ homeassistant/components/supla/const.py
@@ -1,0 +1,6 @@
+"""Constants for SUPLA."""
+
+DOMAIN = "supla"
+
+SUPLA_SERVERS = "supla_servers"
+SUPLA_COORDINATORS = "supla_coordinators"
