+++ homeassistant/components/supla/entity.py
@@ -3,16 +3,21 @@
 from __future__ import annotations
 
 import logging
+from typing import Any
 
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
+
+from .coordinator import SuplaCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
 
-class SuplaEntity(CoordinatorEntity):
+class SuplaEntity(CoordinatorEntity[SuplaCoordinator]):
     """Base class of a SUPLA Channel (an equivalent of HA's Entity)."""
 
-    def __init__(self, config, server, coordinator):
+    def __init__(
+        self, config: dict[str, Any], server: Any, coordinator: SuplaCoordinator
+    ) -> None:
         """Init from config, hookup[ server and coordinator."""
         super().__init__(coordinator)
         self.server_name = config["server_name"]
